renamed text-legibility to adaptable-content
