+++ src/assessments/adaptable-content/assessment.tsx
