+++ src/assessments/adaptable-content/test-steps/high-contrast-mode.tsx
