+++ src/assessments/adaptable-content/test-steps/hover-focus-content.tsx
