+++ src/assessments/adaptable-content/test-steps/reflow.tsx
@@ -2,10 +2,10 @@
 // Licensed under the MIT License.
 import * as React from 'react';
 
-import { TextLegibilityTestStep } from 'assessments/text-legibility/test-steps/test-step';
+import { TextLegibilityTestStep } from 'assessments/adaptable-content/test-steps/test-step';
 import { Requirement } from 'assessments/types/requirement';
 import { link } from 'content/link';
-import * as content from 'content/test/text-legibility/reflow';
+import * as content from 'content/test/adaptable-content/reflow';
 import { ManualTestRecordYourResults } from '../../common/manual-test-record-your-results';
 
 const reflowDescription: JSX.Element = (
