+++ src/assessments/adaptable-content/test-steps/resize-text.tsx
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { TextLegibilityTestStep } from 'assessments/text-legibility/test-steps/test-step';
+import { TextLegibilityTestStep } from 'assessments/adaptable-content/test-steps/test-step';
 import { Requirement } from 'assessments/types/requirement';
 import { link } from 'content/link';
-import * as content from 'content/test/text-legibility/resize-text';
+import * as content from 'content/test/adaptable-content/resize-text';
 import * as React from 'react';
 import { ManualTestRecordYourResults } from '../../common/manual-test-record-your-results';
 import * as Markup from '../../markup';
