+++ src/assessments/assessments.ts
@@ -22,7 +22,7 @@ import { PrerecordedMultimediaAssessment } from './prerecorded-multimedia/assess
 import { RepetitiveContentAssessment } from './repetitive-content/assessment';
 import { SemanticsAssessment } from './semantics/assessment';
 import { SequenceAssessment } from './sequence/assessment';
-import { TextLegibilityAssessment } from './text-legibility/assessment';
+import { TextLegibilityAssessment } from './adaptable-content/assessment';
 import { TimedEventsAssessment } from './timed-events/assessment';
 import { AssessmentsProvider } from './types/assessments-provider';
 import { VisibleFocusOrderAssessment } from './visible-focus-order/assessment';
